+++ components/render/opengl_driver/sources/platform/iphone/gl_iphone_primary_swap_chain.mm
@@ -538,7 +538,7 @@ void PrimarySwapChain::DoneForContext ()
    Получение идентификатора буфера
 */
 
-size_t PrimarySwapChain::GetFrameBufferId ()
+unsigned int PrimarySwapChain::GetFrameBufferId ()
 {
   if (!impl->frame_buffer)
     throw xtl::format_operation_exception ("render::low_level::opengl::iphone::PrimarySwapChain::GetFrameBufferId", "Can't get frame buffer before initializing swap chain");
